Use plain wfHttpError() if short code is not found
Using the MW interface to display an error message on the short domain
itself isn't possible because the normal relative URLs that MW will
generate will all be wrong since we're not on the main domain. Since
users really shouldn't be hitting this anyways, we can use the plain
unstyled wfHttpError() output.

Bug: T116512
Change-Id: I81327fce0930f983d2e258979bd84f4d1d091a7a