Add temporary fix to ensure array is passed to array_map()
We set the default of $wgUrlShortenerDomainsWhitelist to false to
mirror the original, but since it's not yet defined in wmf-config
and we are passing it in array_mpa(), the boolean is causing errors.

When we switch to this varible in wmf-config, array will be correctly
given to the function and there'd be no warnings, but we cannot due
that now since the varibale might not yet be available on all wikis.

Bug: T258056
Change-Id: Iee88850d8cd44cbdbd53ec7f35e6a93f590c5a20
(cherry picked from commit 59488ae48546f22f608875050cb67574f12cc3b4)