Drop newbie RateLimit and make user more strict
for user-global to be able to properly work, we need to remove newbie
and the current user limit is highly permissive, reducing it. We can
change if users complain.

Also setting '&can-bypass' to false to prevent abuse by becoming
admin/bot in a random/test wiki. There's really not much need for
creating this many short urls.

Bug: T267925
Change-Id: I248b7519a783fbd86e3a92b826b0e9b3575bf150