Add link to download QR code
This uses the endroid/qr-code PHP package to generate QR codes.

If the URL is very long ($wgUrlShortenerQrCodeShortenLimit), it is first
shortened before generating a QR code to avoid the image having too high
of density to be scannable. Otherwise, URLs are not shortened so that
once scanned, the visible link is to a domain familiar to the user (as
opposed to w.wiki or whatever  is).

The link is handled by JS, with the no-JS fallback going to
Special:QrCode. This Special page will be added in a subsequent patch.
UrlShortenerUtils::getQrCodeDataUri() is added for use on that page.

Bug: T242467
Depends-On: I8920e0878f676885a795fca8bfd2dc63ec1140cb
Change-Id: Ib3fe1063187389f72047ccbbcd2ec72b843bae17