SpecialQrCode: show error message when QR codes are disabled
The special page shouldn't show a form if $wgUrlShortenerEnableQrCode is
disabled. This commit acts like when $wgUrlShortenerReadOnly is set and
shows an error message instead of the form.

When disabled, the code path means there's no help link to Help:QrCode,
but this should be fine because if the form is hidden, the help page is
irrelevant anyway.

Bug: T356379
Change-Id: I7068d0d11df6ebbab2393b88cfaf7bf2eb8c63b5