Use wgUrlShortenerServer even if wgUrlShortenerTemplate isn't set
In the case where `wgUrlShortenerServer` is set to an actual
wiki or to a URL that is valid, and `wgUrlShortenerTemplate`
isn't set, UrlShortener currently falls back to `wgServer` as the URL,
ignoring whatever the URL shortener server URL was set to.

This patch uses the `wgArticlePath` config setting to formulate a
URL to the `Special:UrlRedirector` special page as the default value
for `wgUrlShortenerTemplate`.

Test plan
=========

A sample config settings in LocalSettings.php like:
```
wfLoadExtension( 'UrlShortener' );
$wgUrlShortenerEnableQrCode = true;
$wgUrlShortenerServer = 'http://foo.wiki';
$wgVirtualDomainsMapping['virtual-urlshortener'] = [ 'db' => 'foowiki' ];
```

This causes requests from `http://bar.wiki` to shorten and render
in the JS dialog with `http://bar.wiki` (as the shortened URL
prefix) instead of `http://foo.wiki` as configured in
`$wgUrlShortenerServer`.

The issue above can be mitigated by changing LocalSettings and
adding a URL template explicitly:

```
wfLoadExtension( 'UrlShortener' );
$wgUrlShortenerEnableQrCode = true;
$wgUrlShortenerServer = 'http://foo.wiki';
$wgUrlShortenerTemplate = '/w/index.php/r/$1';
$wgVirtualDomainsMapping['virtual-urlshortener'] = [ 'db' => 'foowiki' ];
```

The above works correctly since `wgUrlShortenerTemplate` is set
and that code path appends the short URL to the short server URL.

This patch ensures that with or without a URL template config set,
URL shortening works consistently and cleanly.

Bug: T413069
Change-Id: I8d98fc816bee54bc7ed8318e11be277a04f655ab